requirements.txt
numpy==2.1.3
pandas==2.2.3
matplotlib==3.9.2
joblib==1.4.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
kaggle==1.6.17
